requirements/dev.txt
-r base.txt
pytest>=8.0
pytest-django>=4.9
pytest-cov>=6.0
factory-boy>=3.3
faker>=30.0
django-debug-toolbar>=4.4
ruff>=0.8
black>=24.0
